backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic>=2.7.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
stripe==9.11.0
celery==5.3.6
redis==5.0.4
httpx==0.27.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1
bs4==0.0.2
pydantic-settings>=2.0.0